Prepare for finish
This will use the released versions of the skrillax-network crates as well as make it possible to cancel properly. Additionally, this will now use rustls in build (we don't actually need https anyway) in order to avoid openssl dependency. To bring it all together, this now provides a Dockerfile to ship it.